requirements.txt
pandas
scikit-learn
huggingface_hub
streamlit
joblib
numpy
mlflow
pyyaml
python-dotenv
